mi_proyecto_mlops/airflow/requirements.txt
# # Requerimientos para el entorno de Airflow
# mlflow
# scikit-learn
# great-expectations
# lightgbm
# # --- Nuevas librerías para el Bot de Trading ---
# xgboost
# optuna
# TA-Lib
# python-binance


# Requerimientos para el entorno de Airflow (con versiones fijadas)
mlflow==2.8.0
scikit-learn==1.3.2
great-expectations
lightgbm
xgboost==2.0.0
optuna==3.4.0
TA-Lib
python-binance
zeep
xlrd